chore: move PermitProviderConnector to utils
